pipeline/requirements.txt
beautifulsoup4==4.11.1
Flask==2.1.1
google-api-core==2.7.2
google-auth==2.6.6
google-cloud==0.34.0
google-cloud-core==2.3.0
google-cloud-pubsub==2.12.0
lxml==4.8.0
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
pytest==7.4.3
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
sdmxthon==2.5
selenium==4.15.2
Werkzeug==2.1.1
xlrd==2.0.1